requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.0.0
scikit-learn>=0.24.0
torch>=1.9.0
torchvision>=0.10.0
streamlit>=1.12.0
kagglehub>=0.2.0